bus_demux: Remove selected_slv
selected_slv is a natural whose range after syntesis is a power of 2.
This means that if you have, for example, 3 slaves it is possible for
selected_slv to be set to a value which is out of array bounds. To fix
this, we remove this variable and set the output directly in the loop.